Break up IsGraphQLOutputType.scala
